+++ perf/ListPool.Benchmarks/Enumerable_ToList.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Order;
 
@@ -9,27 +10,23 @@ namespace ListPool.Benchmarks
     [MemoryDiagnoser]
     [GcServer(true)]
     [GcConcurrent]
-    public class ListPoolCreateBenchmarks
+    public class Enumerable_ToList
     {
+        private IEnumerable<int> _items => Enumerable.Range(0, N);
+
         [Params(100, 1_000, 10_000)]
         public int N { get; set; }
-
-        [Benchmark(Baseline = true)]
-        public void List()
-        {
-            _ = new List<int>(N);
-        }
 
         [Benchmark]
         public void ListPool()
         {
-            using ListPool<int> list = new ListPool<int>(N);
+            using ListPool<int> _ = _items.ToListPool();
         }
 
-        [Benchmark]
-        public void ValueListPool()
+        [Benchmark(Baseline = true)]
+        public void Linq()
         {
-            using ValueListPool<int> list = new ValueListPool<int>(N);
+            _ = _items.ToList();
         }
     }
 }
