+++ perf/ListPool.Benchmarks/List_RemoveAt.cs
@@ -9,7 +9,7 @@ namespace ListPool.Benchmarks
     [MemoryDiagnoser]
     [GcServer(true)]
     [GcConcurrent]
-    public class ListPoolContainsBenchmark
+    public class List_RemoveAt
     {
         private List<int> _list;
         private ListPool<int> _listPool;
@@ -17,8 +17,8 @@ namespace ListPool.Benchmarks
         [Params(100, 1_000, 10_000)]
         public int N { get; set; }
 
-        [GlobalSetup]
-        public void GlobalSetup()
+        [IterationSetup]
+        public void IterationSetup()
         {
             _list = new List<int>(N);
             _listPool = new ListPool<int>(N);
@@ -30,22 +30,22 @@ namespace ListPool.Benchmarks
             }
         }
 
-        [GlobalCleanup]
-        public void GlobalCleanup()
+        [IterationCleanup]
+        public void IterationCleanup()
         {
             _listPool.Dispose();
         }
 
         [Benchmark(Baseline = true)]
-        public bool List()
+        public void List()
         {
-            return _list.Contains(N / 2);
+            _list.RemoveAt(N / 2);
         }
 
         [Benchmark]
-        public bool ListPool()
+        public void ListPool()
         {
-            return _listPool.Contains(N / 2);
+            _listPool.RemoveAt(N / 2);
         }
     }
 }
