+++ perf/ListPool.Benchmarks/Utf8Json/Formatters/Models/DummyClass.cs
@@ -1,0 +1,7 @@
+namespace ListPool.Benchmarks.Formatters.Utf8Json
+{
+    public class DummyClass
+    {
+        public string Text { get; set; }
+    }
+}
