+++ perf/ListPool.Benchmarks/Utf8Json/Formatters/Models/DummyClassUsingListPool.cs
@@ -1,0 +1,14 @@
+using System;
+using ListPool.Formatters.Utf8Json;
+using Utf8Json;
+
+namespace ListPool.Benchmarks.Formatters.Utf8Json
+{
+    public class DummyClassUsingListPool : IDisposable
+    {
+        [JsonFormatter(typeof(StringAsListPoolOfCharsFormatter))]
+        public ListPool<char> Text { get; set; }
+
+        public void Dispose() => Text?.Dispose();
+    }
+}
