+++ perf/ListPool.Benchmarks/Utf8Json/Formatters/StringAsListPoolOfChars_Deserialize.cs
@@ -1,0 +1,39 @@
+using System.Text;
+using BenchmarkDotNet.Attributes;
+using Utf8Json;
+
+namespace ListPool.Benchmarks.Formatters.Utf8Json
+{
+    [MemoryDiagnoser]
+    public class StringAsListPoolOfChars_Deserialize
+    {
+        private byte[] _json;
+
+        [Params(500, 10_000)] public int N { get; set; }
+
+        [GlobalSetup]
+        public void GlobalSetup()
+        {
+            StringBuilder sb = new StringBuilder(N + 10);
+
+            sb.Append("{\"Text\":\"");
+            for (int i = 0; i < N; i++)
+            {
+                sb.Append("a");
+            }
+            sb.Append("\"}");
+
+            _json = Encoding.UTF8.GetBytes(sb.ToString());
+        }
+
+        [Benchmark(Baseline = true)]
+        public int String() => JsonSerializer.Deserialize<DummyClass>(_json).Text.Length;
+
+        [Benchmark]
+        public int ListPool()
+        {
+            using var dummyClass = JsonSerializer.Deserialize<DummyClassUsingListPool>(_json);
+            return dummyClass.Text.Count;
+        }
+    }
+}
